include: moved RBD's generic FunctionContext wrapper
FunctionContext uses boost::function to allow the Context
interface to invoke arbitrary functions.  This eliminates
the need to create derived Context classes for every use
case.
